Add the parser and basic testing
This is a gigantic change, because, well, making a parser isn't
something that you can really split up. Dependencies had to be added,
configs had to be changed too. This should be ready for merging, and
then the CLI can use this. Maybe add some utility functions to make it
simpler and clearer.